+++ include/boost/beast/_experimental/http/icy_stream.hpp
@@ -224,11 +224,14 @@ public:
     */
     template<
         class MutableBufferSequence,
-        class ReadHandler>
+        class ReadHandler =
+            net::default_completion_token_t<executor_type>
+    >
     BOOST_BEAST_ASYNC_RESULT2(ReadHandler)
     async_read_some(
         MutableBufferSequence const& buffers,
-        ReadHandler&& handler);
+        ReadHandler&& handler =
+            net::default_completion_token_t<executor_type>{});
 
     /** Write some data to the stream.
     
@@ -303,11 +306,14 @@ public:
     */
     template<
         class ConstBufferSequence,
-        class WriteHandler>
+        class WriteHandler =
+            net::default_completion_token_t<executor_type>
+    >
     BOOST_BEAST_ASYNC_RESULT2(WriteHandler)
     async_write_some(
         ConstBufferSequence const& buffers,
-        WriteHandler&& handler);
+        WriteHandler&& handler =
+            net::default_completion_token_t<executor_type>{});
 };
 
 } // http