+++ include/boost/beast/core/detect_ssl.hpp
@@ -316,7 +316,9 @@ detect_ssl(
 template<
     class AsyncReadStream,
     class DynamicBuffer,
-    class CompletionToken>
+    class CompletionToken =
+        net::default_completion_token_t<beast::executor_type<AsyncReadStream>>
+>
 #if BOOST_BEAST_DOXYGEN
 BOOST_ASIO_INITFN_RESULT_TYPE(CompletionToken, void(error_code, bool))
 #else
@@ -325,7 +327,8 @@ auto
 async_detect_ssl(
     AsyncReadStream& stream,
     DynamicBuffer& buffer,
-    CompletionToken&& token) ->
+    CompletionToken&& token = net::default_completion_token_t<
+            beast::executor_type<AsyncReadStream>>{}) ->
         typename net::async_result<
             typename std::decay<CompletionToken>::type, /*< `async_result` customizes the return value based on the completion token >*/
             void(error_code, bool)>::return_type; /*< This is the signature for the completion handler >*/