+++ test/doc/core_4_layers.cpp
@@ -243,11 +243,15 @@ public:
     }
 
     /// Read some data from the stream asynchronously
-    template <class MutableBufferSequence, class ReadHandler>
+    template<
+        class MutableBufferSequence,
+        class ReadHandler =
+            net::default_completion_token_t<executor_type>>
     BOOST_BEAST_ASYNC_RESULT2(ReadHandler)
     async_read_some(
         MutableBufferSequence const& buffers,
-        ReadHandler&& handler)
+        ReadHandler&& handler =
+            net::default_completion_token_t<executor_type>{})
     {
         return net::async_initiate<
             ReadHandler,
@@ -259,11 +263,16 @@ public:
     }
 
     /// Write some data to the stream asynchronously
-    template <class ConstBufferSequence, class WriteHandler>
+    template<
+        class ConstBufferSequence,
+        class WriteHandler =
+            net::default_completion_token_t<executor_type>>
     BOOST_BEAST_ASYNC_RESULT2(WriteHandler)
     async_write_some(
         ConstBufferSequence const& buffers,
-        WriteHandler&& handler)
+        WriteHandler&& handler =
+            net::default_completion_token_t<
+                executor_type>{})
     {
         return net::async_initiate<
             WriteHandler,